% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq.R
\name{analyzeSalmonDataDeSeq2}
\alias{analyzeSalmonDataDeSeq2}
\title{Analyze Salmon quant data using DeSeq2}
\usage{
analyzeSalmonDataDeSeq2(
  salmon,
  outdir,
  exclude_samples = c(),
  protein_only = FALSE,
  groupcol = "group",
  levels = c("Control", "Case")
)
}
\arguments{
\item{levels}{}
}
\value{

}
\description{
Analyze Salmon quant data using DeSeq2
}
