% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq.R
\name{analyzeSalmonDeSeq2}
\alias{analyzeSalmonDeSeq2}
\title{Analyze Salmon quant data using DeSeq2}
\usage{
analyzeSalmonDeSeq2(project, txi, groupcol, outdir, protein_only = FALSE)
}
\arguments{
\item{protein_only}{}
}
\value{

}
\description{
Analyze Salmon quant data using DeSeq2
}
