% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframes.R
\name{appendValues}
\alias{appendValues}
\title{Append values to vector}
\usage{
appendValues(values1, values2)
}
\arguments{
\item{values2}{}
}
\value{

}
\description{
Append values to vector
}
\examples{
appendValues('a,b,c,d,e,f,g','h,i,j,k,l')
appendValues(c(),'q')
}
