% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframes.R
\name{checkColExists}
\alias{checkColExists}
\title{Throws an error if a column does not exist in a dataframe}
\usage{
checkColExists(data, colname)
}
\arguments{
\item{colname}{}
}
\value{

}
\description{
Throws an error if a column does not exist in a dataframe
}
