% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframes.R
\name{colExists}
\alias{colExists}
\title{Checks if a column exists in a dataframe}
\usage{
colExists(data, colname)
}
\arguments{
\item{colname}{}
}
\value{

}
\description{
Checks if a column exists in a dataframe
}
