% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{concat}
\alias{concat}
\title{Concatenate strings similar to paste0}
\usage{
concat(..., sep = "", collapse = NULL)
}
\arguments{
\item{...}{strings to concatenate}

\item{sep}{separator, defaults to empty string}

\item{collapse}{defaults to NULL}
}
\value{
concatenated string
}
\description{
Concatenate strings similar to paste0
}
\examples{
concat('abc','def','ghi')
}
