% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframes.R
\name{containsElement}
\alias{containsElement}
\title{Checks if a vector contains a value}
\usage{
containsElement(x, value)
}
\arguments{
\item{value}{}
}
\value{

}
\description{
Checks if a vector contains a value
}
