% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq2.R
\name{DeSeq2Class}
\alias{DeSeq2Class}
\title{R6 class for DESeq2 analysis}
\value{

}
\description{
R6 class for DESeq2 analysis

R6 class for DESeq2 analysis
}
\examples{
project <- DeSeq2Class$new(salmon, groupcol, outdir)
createWgcnaClass


## ------------------------------------------------
## Method `DeSeq2Class$createWgcnaClassFromDeseq2`
## ------------------------------------------------

wgcna <- createWgcnaClassFromDeseq2(vsd, res_lfc, datTraits, degenes_only=FALSE, ngenes=5000)
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DeSeq2Class$new()}}
\item \href{#method-getResults}{\code{DeSeq2Class$getResults()}}
\item \href{#method-makeTable}{\code{DeSeq2Class$makeTable()}}
\item \href{#method-createWgcnaClassFromDeseq2}{\code{DeSeq2Class$createWgcnaClassFromDeseq2()}}
\item \href{#method-clone}{\code{DeSeq2Class$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeSeq2Class$new(dds, outdir)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{outdir}}{directory to write DGE results to}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getResults"></a>}}
\if{latex}{\out{\hypertarget{method-getResults}{}}}
\subsection{Method \code{getResults()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeSeq2Class$getResults(p = NULL, padj = 0.05, logfc = 1.5, n = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-makeTable"></a>}}
\if{latex}{\out{\hypertarget{method-makeTable}{}}}
\subsection{Method \code{makeTable()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeSeq2Class$makeTable(num = 50, padj = 1, lfc = 1.5)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-createWgcnaClassFromDeseq2"></a>}}
\if{latex}{\out{\hypertarget{method-createWgcnaClassFromDeseq2}{}}}
\subsection{Method \code{createWgcnaClassFromDeseq2()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeSeq2Class$createWgcnaClassFromDeseq2(
  datTraits,
  p = NULL,
  padj = NULL,
  logfc = NULL,
  n = 5000
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{datTraits}}{}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{wgcna <- createWgcnaClassFromDeseq2(vsd, res_lfc, datTraits, degenes_only=FALSE, ngenes=5000)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeSeq2Class$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
