% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{getCountsByGroup}
\alias{getCountsByGroup}
\title{getCountsByGroup adjusts the counts and samples to exclude missing samples}
\usage{
getCountsByGroup(counts, project, groupcol)
}
\arguments{
\item{counts}{RNAseq count data}

\item{project}{project data containing sample and group information}

\item{groupcol}{group column}
}
\value{

}
\description{
getCountsByGroup adjusts the counts and samples to exclude missing samples
}
