% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{getGroupLevels}
\alias{getGroupLevels}
\title{getGroupLevels returns uniquq ordered levels for specified group}
\usage{
getGroupLevels(project, groupcol)
}
\arguments{
\item{project}{project object}

\item{groupcol}{group column}
}
\value{

}
\description{
getGroupLevels returns uniquq ordered levels for specified group
}
\examples{
getGroupLevels(project, groupcol='default')
}
