% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlsgr.R
\docType{package}
\name{hlsgr}
\alias{hlsgr}
\title{hlsgr: bioinformatics and genomics utilities for HLSG}
\description{
The package contains bioinformatics and genomics utilities for HLSG
}
\section{Foo functions}{

The foo functions ...
}

