% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolomics.R
\name{HmtMetabolomicsDataClass}
\alias{HmtMetabolomicsDataClass}
\title{R6 class for HMT metabolomics data}
\value{

}
\description{
R6 class for HMT metabolomics data

R6 class for HMT metabolomics data
}
\examples{
project <- HmtMetabolomicsDataClass$new(paste0(metabolomicsdir, '/metabolome_normalization.txt'))
}
\section{Super classes}{
\code{\link[hlsgr:AbstractBaseClass]{hlsgr::AbstractBaseClass}} -> \code{\link[hlsgr:AbstractMetabolomicsDataClass]{hlsgr::AbstractMetabolomicsDataClass}} -> \code{HmtMetabolomicsDataClass}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{HmtMetabolomicsDataClass$new()}}
\item \href{#method-clone}{\code{HmtMetabolomicsDataClass$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HmtMetabolomicsDataClass$new(filename)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{filename}}{filename}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HmtMetabolomicsDataClass$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
