% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframes.R
\name{loadDataFrame}
\alias{loadDataFrame}
\title{Load data frame}
\usage{
loadDataFrame(
  filename,
  idcol = NULL,
  header = TRUE,
  stringsAsFactors = FALSE,
  check.names = TRUE,
  comment.char = "#",
  sep = "\\t",
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Load data frame
}
