% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq.R
\name{loadSalmonData}
\alias{loadSalmonData}
\title{Load Salmon quant files}
\usage{
loadSalmonData(samples, quantdir)
}
\arguments{
\item{quantdir}{}
}
\value{

}
\description{
Load Salmon quant files
}
\examples{
salmon <- loadSalmonData(samples, quantdir)
}
