% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq.R
\name{loadSalmonSamples}
\alias{loadSalmonSamples}
\title{Load sample information for Salmon-mapped samples}
\usage{
loadSalmonSamples(filename)
}
\arguments{
\item{filename}{}
}
\value{

}
\description{
Load sample information for Salmon-mapped samples
}
\examples{
samples <- loadSalmonSamples()
}
