% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{loadSamples}
\alias{loadSamples}
\title{Load sample information}
\usage{
loadSamples(filename)
}
\arguments{
\item{countdir}{}
}
\value{

}
\description{
Load sample information
}
\examples{
filename <- 'samples.txt'
samples <- loadSamples(filename)
}
