% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{loadSubjects}
\alias{loadSubjects}
\title{loadSubjects}
\usage{
loadSubjects(filename)
}
\arguments{
\item{filename}{}
}
\value{

}
\description{
loadSubjects
}
\examples{
subjects <- loadSubjects('subjects.txt')
}
