% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq.R
\name{loadSubreadData}
\alias{loadSubreadData}
\title{Load Salmon quant files}
\usage{
loadSubreadData(samples, countdir)
}
\arguments{
\item{countdir}{}
}
\value{

}
\description{
Load Salmon quant files
}
\examples{
countdir <-
subread <- loadSubreadData(samples, countdir)
}
