% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq.R
\name{lookupGenes}
\alias{lookupGenes}
\title{Return a vector gene names corresponding to a vector of Ensembl IDs}
\usage{
lookupGenes(ensembl_ids)
}
\arguments{
\item{ensembl_ids}{vector of ensembl IDs}
}
\value{

}
\description{
Return a vector gene names corresponding to a vector of Ensembl IDs
}
\examples{
lookupGenes(c('ENSG00000211459', 'ENSG00000211958 '))
}
