% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq.R
\name{makeDeSeq2Table}
\alias{makeDeSeq2Table}
\title{Create a formatted table of differentially regulated genes}
\usage{
makeDeSeq2Table(result, num = 50, padj = 1, lfc = 1.5)
}
\arguments{
\item{result}{DeSeq2 analysis results}

\item{num}{max results to show}

\item{padj}{cutoff value for adjusted p-value}

\item{lfc}{log fold change cutoff (positive or negative)}
}
\value{

}
\description{
Create a formatted table of differentially regulated genes
}
\examples{
makeDeSeq2Table(result)
}
