% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_function.R
\name{makeMeSmile}
\alias{makeMeSmile}
\title{Reassure A User That They Are Doing Well}
\usage{
makeMeSmile(me = "I")
}
\arguments{
\item{me}{the user to reassure}
}
\value{
NULL (invisibly). Used for the side-effect of generating a
\code{message}.
}
\description{
Look at that, they're building a documented function already!
}
\examples{
makeMeSmile("You")
}
