% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome.R
\name{plotChromosomes}
\alias{plotChromosomes}
\title{Create a circular chromsome plot showing location of variants}
\usage{
plotChromosomes(snv = character(0), cnv = character(0), fusions = character(0))
}
\arguments{
\item{fusions}{}
}
\value{

}
\description{
https://www.bioconductor.org/packages/release/bioc/vignettes/ggbio/inst/doc/ggbio.pdf
}
\examples{
p <- hlsgr::plotChromosomes(cnv=c('GNA11', 'KRAS'), snv=c('PIK3CA', 'APC', 'FGFR2'), fusions=c('ALK', 'NBR1'))
}
