% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq.R
\name{plotDeSeq2Heatmap}
\alias{plotDeSeq2Heatmap}
\title{Create a heatmap for DeSeq2 RNASeq data}
\usage{
plotDeSeq2Heatmap(
  result,
  num = 50,
  fontsize = 6,
  color = colorRampPalette(c("blue", "white", "red"))(50)
)
}
\arguments{
\item{color}{}
}
\value{

}
\description{
Create a heatmap for DeSeq2 RNASeq data
}
\examples{
plotHeatmap(result, 10)
}
