% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kegg.R
\name{plotKeggPathway}
\alias{plotKeggPathway}
\title{Title}
\usage{
plotKeggPathway(pathway, genes, kegg_dir, bg_col = "firebrick1")
}
\arguments{
\item{bg_col}{}
}
\value{

}
\description{
Title
}
\examples{
hlsgr::plotKeggPathway(pathway='hsa05200', genes=c(672, 811, 2065, 2260, 201163, 2321, 6416, 5727, 5888, 5925, 7157), kegg_dir='c:/workspace/hlsg/data/kegg')
}
