% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq.R
\name{readSalmonData}
\alias{readSalmonData}
\title{Read Salmon data from RDS file}
\usage{
readSalmonData(filename)
}
\arguments{
\item{filename}{.rds file holding salmon quant data}
}
\value{

}
\description{
Read Salmon data from RDS file
}
\examples{
salmon = readSalmonData('salmon.rds')
}
