% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq.R
\name{saveSalmonData}
\alias{saveSalmonData}
\title{Write Salmon quant data to to rds file for faster loading across sessions}
\usage{
saveSalmonData(salmon, filename)
}
\arguments{
\item{salmon}{salmon quant data}

\item{filename}{.rds file to store quant data in}
}
\value{

}
\description{
Write Salmon quant data to to rds file for faster loading across sessions
}
\examples{
saveSalmonData(salmon, 'salmon.rds')
}
