% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframes.R
\name{splitFields}
\alias{splitFields}
\title{Split fields on delimiter}
\usage{
splitFields(str, delimiter = ",")
}
\arguments{
\item{delimiter}{}
}
\value{

}
\description{
Split fields on delimiter
}
\examples{
splitFields('abc, def, ghi')
}
