% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome.R
\name{splitGenes}
\alias{splitGenes}
\title{Split list of genes into a vector}
\usage{
splitGenes(values = character(0))
}
\arguments{
\item{values}{}
}
\value{

}
\description{
Splits comma-delimited list and returns only genes found in GENESYMBOL list
}
\examples{
hlsgr::splitGenes('ALK,NR1')#
hlsgr::splitGenes(c('PIK3CA', 'APC', 'FGFR2'))
}
