% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq.R
\name{subsetTxi}
\alias{subsetTxi}
\title{Subset txi data by sample or gene}
\usage{
subsetTxi(txi, samples, include_genes = rownames(txi$counts))
}
\arguments{
\item{include_genes}{}
}
\value{

}
\description{
Subset txi data by sample or gene
}
