% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolomics.R
\name{TakedaMetabolomicsDataClass}
\alias{TakedaMetabolomicsDataClass}
\title{R6 class for Takeda metabolomics data}
\value{

}
\description{
R6 class for Takeda metabolomics data

R6 class for Takeda metabolomics data
}
\examples{
project <- TakedaMetabolomicsDataClass$new()
}
\section{Super classes}{
\code{\link[hlsgr:AbstractBaseClass]{hlsgr::AbstractBaseClass}} -> \code{\link[hlsgr:AbstractMetabolomicsDataClass]{hlsgr::AbstractMetabolomicsDataClass}} -> \code{TakedaMetabolomicsDataClass}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{TakedaMetabolomicsDataClass$new()}}
\item \href{#method-clone}{\code{TakedaMetabolomicsDataClass$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TakedaMetabolomicsDataClass$new(filename)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{filename}}{filename}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TakedaMetabolomicsDataClass$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
