% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{trim}
\alias{trim}
\title{Trims whitespace from string}
\usage{
trim(str)
}
\arguments{
\item{str}{input string to trim}
}
\value{
trimmed string
}
\description{
Trims whitespace from string
}
\examples{
trim(' 	 abc def 	 ')
}
