% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframes.R
\name{writeTable}
\alias{writeTable}
\title{Write table to file}
\usage{
writeTable(
  table,
  filename,
  verbose = TRUE,
  row.names = FALSE,
  col.names = TRUE,
  eol = "\\n",
  na = ""
)
}
\arguments{
\item{na}{}
}
\value{

}
\description{
http://r.789695.n4.nabble.com/Appending-strings-at-the-beginning-of-a-text-file-td901370.html
}
