% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{getSamplesByGroup}
\alias{getSamplesByGroup}
\title{getSamplesByGroup}
\usage{
getSamplesByGroup(project, groupcol, excluded = c())
}
\arguments{
\item{groupcol}{group column}

\item{excluded}{vector of samples to exclude}

\item{samples}{loaded from samples.txt}

\item{groups}{loaded from groups.txt}
}
\value{

}
\description{
getSamplesByGroup
}
