% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{loadSubreadData}
\alias{loadSubreadData}
\title{Load Subread count files}
\usage{
loadSubreadData(project, countdir)
}
\arguments{
\item{countdir}{}
}
\value{

}
\description{
Load Subread count files
}
\examples{
countdir <-
subread <- loadSubreadData(samples, countdir)
}
