% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{NgsProjectClass}
\alias{NgsProjectClass}
\title{R6 class for project information}
\value{

}
\description{
R6 class for project information

R6 class for project information
}
\examples{
project <- NgsProjectClass$new(projectdir)
getGroupLevels returns unique ordered levels for specified group


## ------------------------------------------------
## Method `NgsProjectClass$getGroupLevels`
## ------------------------------------------------

getGroupLevels(project, groupcol='default')
getSamplesByGroup

}
\section{Super class}{
\code{\link[hlsgr:AbstractBaseClass]{hlsgr::AbstractBaseClass}} -> \code{NgsProjectClass}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{NgsProjectClass$new()}}
\item \href{#method-getGroupLevels}{\code{NgsProjectClass$getGroupLevels()}}
\item \href{#method-getSamplesByGroup}{\code{NgsProjectClass$getSamplesByGroup()}}
\item \href{#method-getCountsByGroup}{\code{NgsProjectClass$getCountsByGroup()}}
\item \href{#method-getDatTraits}{\code{NgsProjectClass$getDatTraits()}}
\item \href{#method-createDeseq2ClassFromCounts}{\code{NgsProjectClass$createDeseq2ClassFromCounts()}}
\item \href{#method-createDeseq2ClassFromSalmon}{\code{NgsProjectClass$createDeseq2ClassFromSalmon()}}
\item \href{#method-createEdgeRClassFromCounts}{\code{NgsProjectClass$createEdgeRClassFromCounts()}}
\item \href{#method-createEdgeRClassFromSalmon}{\code{NgsProjectClass$createEdgeRClassFromSalmon()}}
\item \href{#method-createLimmaVoomClassFromCounts}{\code{NgsProjectClass$createLimmaVoomClassFromCounts()}}
\item \href{#method-createLimmaVoomClassFromSalmon}{\code{NgsProjectClass$createLimmaVoomClassFromSalmon()}}
\item \href{#method-clone}{\code{NgsProjectClass$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NgsProjectClass$new(dir)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dir}}{folder containing project and sample information}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getGroupLevels"></a>}}
\if{latex}{\out{\hypertarget{method-getGroupLevels}{}}}
\subsection{Method \code{getGroupLevels()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NgsProjectClass$getGroupLevels(groupcol)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{groupcol}}{group column}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{

}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{getGroupLevels(project, groupcol='default')
getSamplesByGroup

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getSamplesByGroup"></a>}}
\if{latex}{\out{\hypertarget{method-getSamplesByGroup}{}}}
\subsection{Method \code{getSamplesByGroup()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NgsProjectClass$getSamplesByGroup(groupcol)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{groupcol}}{group column}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{

}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getCountsByGroup"></a>}}
\if{latex}{\out{\hypertarget{method-getCountsByGroup}{}}}
\subsection{Method \code{getCountsByGroup()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NgsProjectClass$getCountsByGroup(counts, groupcol)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getDatTraits"></a>}}
\if{latex}{\out{\hypertarget{method-getDatTraits}{}}}
\subsection{Method \code{getDatTraits()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NgsProjectClass$getDatTraits(cols)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-createDeseq2ClassFromCounts"></a>}}
\if{latex}{\out{\hypertarget{method-createDeseq2ClassFromCounts}{}}}
\subsection{Method \code{createDeseq2ClassFromCounts()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NgsProjectClass$createDeseq2ClassFromCounts(
  rnaseq,
  groupcol,
  outdir,
  identifier = "ensembl_id"
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-createDeseq2ClassFromSalmon"></a>}}
\if{latex}{\out{\hypertarget{method-createDeseq2ClassFromSalmon}{}}}
\subsection{Method \code{createDeseq2ClassFromSalmon()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NgsProjectClass$createDeseq2ClassFromSalmon(salmon, groupcol, outdir)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-createEdgeRClassFromCounts"></a>}}
\if{latex}{\out{\hypertarget{method-createEdgeRClassFromCounts}{}}}
\subsection{Method \code{createEdgeRClassFromCounts()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NgsProjectClass$createEdgeRClassFromCounts(
  rnaseq,
  groupcol,
  outdir,
  identifier = "ensembl_id"
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-createEdgeRClassFromSalmon"></a>}}
\if{latex}{\out{\hypertarget{method-createEdgeRClassFromSalmon}{}}}
\subsection{Method \code{createEdgeRClassFromSalmon()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NgsProjectClass$createEdgeRClassFromSalmon(salmon, groupcol, outdir)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-createLimmaVoomClassFromCounts"></a>}}
\if{latex}{\out{\hypertarget{method-createLimmaVoomClassFromCounts}{}}}
\subsection{Method \code{createLimmaVoomClassFromCounts()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NgsProjectClass$createLimmaVoomClassFromCounts(
  rnaseq,
  groupcol,
  outdir,
  identifier = "ensembl_id"
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-createLimmaVoomClassFromSalmon"></a>}}
\if{latex}{\out{\hypertarget{method-createLimmaVoomClassFromSalmon}{}}}
\subsection{Method \code{createLimmaVoomClassFromSalmon()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NgsProjectClass$createLimmaVoomClassFromSalmon(salmon, groupcol, outdir)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NgsProjectClass$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
